ARM: dts: odroidxu: enable USB 3503a hub
The Odroid has a SMSC USB3503 connected in hardware only mode like a
PHY, support it using the usb-nop-xceiv binding.  Note that due to a
regrettable decision to use a regulator to represent the reset signal
this uses a fixed voltage regulator to do that, there is a plan to use
the reset controller binding once that is merged so it does not seem
worthwhile to fix the usb-nop-xceiv driver at this point.